Progress towards Choices as Radio Buttons
Note that this displays right, but does *not* yet work!  The problem is
that the radio buttons aren't wired to be sensitive, the div around them
is, and when the div tries to send to the server, it doesn't know to
send the currently-selected button.

This all needs a lot of rethinking, sadly.  It might be the point that
demands a true rethink of the live controls...